fix TTI scenario, runthematrix updtaed accordingly , customs standardized
